Removed superfluous conversions no required thanks to auto unboxing. Using more try-catch with resource statements. Removed unused imports.

+++ src/main/java/com/santiagolizardo/beobachter/gui/dialogs/components/EditionPanel.java
@@ -296,7 +296,7 @@ public class EditionPanel extends JPanel {
 
 	public void saveChanges() {
 		short interval = Short
-				.valueOf(spnRefresh.getValue().toString()).shortValue();
+				.valueOf(spnRefresh.getValue().toString());
 		logType.setRefreshInterval(interval);
 		logType.setRules(modelRules.getRules());
